backend/requirements.txt
torch==2.10
pandas==2.2.2
nltk==3.9.1
scikit-learn==1.6.1
filelock==3.25.2
typing-extensions==4.15.0
setuptools==75.2.0
sympy==1.14.0
networkx==3.6.1
jinja2==3.1.6
fsspec==2025.3.0
numpy==2.0.2
python-dateutil==2.9.0
pytz==2025.2
tzdata==2025.3
click==8.3.1
joblib==1.5.3
regex==2025.11.3
tqdm==4.67.3
scipy==1.16.3
threadpoolctl==3.6.0
six==1.17.0
mpmath==1.3.0
MarkupSafe==3.0.3
django
djangorestframework
django-cors-headers
python-dotenv
firebase-admin
openai
google-generativeai
# ── LangChain (for multi-agent pipeline) ──
langchain
langchain-openai
langchain-core